fixing build

+++ plugins/writeopia_serialization/src/commonMain/kotlin/io/writeopia/sdk/serialization/data/IconApi.kt
@@ -3,4 +3,4 @@ package io.writeopia.sdk.serialization.data
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class IconApi(val label: String, val tint: Long?)
+data class IconApi(val label: String, val tint: Int?)
